Изменен дизайн приложения, добавлено редактирование аккаунта, добавлена тех. поддержка, изменино создание задачи, изменина логика BarCodeAnalyser и тд.

+++ app/src/main/java/com/rastilka/domain/use_case/GetPointsUseCase.kt
@@ -4,6 +4,6 @@ import com.rastilka.domain.repository.MainRepository
 import javax.inject.Inject
 
 class GetPointsUseCase @Inject constructor(private val repository: MainRepository) {
-    suspend operator fun invoke(fromUserId: String, points: Int, comment: String? = null) =
+    suspend operator fun invoke(fromUserId: String, points: Long, comment: String? = null) =
         repository.getPoint(fromUserId, points, comment)
 }
