+++ app/src/main/java/com/rastilka/domain/use_case/SendPointsUseCase.kt
@@ -4,6 +4,6 @@ import com.rastilka.domain.repository.MainRepository
 import javax.inject.Inject
 
 class SendPointsUseCase @Inject constructor(private val repository: MainRepository) {
-    suspend operator fun invoke(toUserId: String, points: Int, comment: String? = null) =
+    suspend operator fun invoke(toUserId: String, points: Long, comment: String? = null) =
         repository.sendPoint(toUserId, points, comment)
 }
